pyproject.toml
[tool.poetry]
name = "airbnb-clone-backend"
version = "0.1.0"
description = ""
authors = ["simon toy project"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
pillow = "^10.0.1"
djangorestframework = "^3.14.0"
pyjwt = "^2.8.0"
django-environ = "^0.11.2"
django-cors-headers = "^4.3.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
